requirements.txt
djangorestframework
transformers
PyPDF2
torch
django
django-cors-headers
gunicorn
django-rest-swagger
pymupdf